pyproject.toml
[tool.poetry]
name = "rcmreloaded"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
atproto = "^0.0.65"
pillow = "^12.1.0"
pydantic = "^2.12.5"
dotenv = "^0.9.9"
numpy = "^2.4.1"
scikit-learn = "^1.8.0"
scipy = "^1.17.0"
ipykernel = "^7.1.0"
colorspacious = "^1.1.2"
matplotlib = "^3.10.8"
coloraide = "^8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
